+++ date-times.d.ts
@@ -1,3 +1,5 @@
-﻿
+
+/** Unix epoch UTC millisecond timestamp
+ */
 declare interface TimestampUtc {
 }