+++ test/DateTimesTest.ts
@@ -5,6 +5,8 @@ import chai = require("chai");
 import mocha = require("mocha");
 import DateTimes = require("../DateTimes");
 
+var { DateConstants, Dates, Timestamps, DotNetJsonDate } = DateTimes;
+
 var asr = chai.assert;
 
 
@@ -12,22 +14,22 @@ suite("DateTimes", function DateTimesTest() {
 
     function testDotNetJson(asr: Chai.AssertStatic) {
         var date = new Date(2000, 0, 27);
-        var jsonTimestamp = DateTimes.DotNetJsonDate.toDotNetJson(date);
-        var res = DateTimes.DotNetJsonDate.parseDotNetJson(jsonTimestamp);
+        var jsonTimestamp = DotNetJsonDate.toDotNetJson(date);
+        var res = DotNetJsonDate.parseDotNetJson(jsonTimestamp);
         asr.equal(date.getTime(), res.getTime());
     }
 
 
     function testDotNetJsonTimestamp(asr: Chai.AssertStatic) {
         var now = Date.now();
-        var jsonTimestamp = DateTimes.Timestamp.toDotNetJson(now, false);
-        var res = DateTimes.Timestamp.parseDotNetJson(jsonTimestamp);
+        var jsonTimestamp = Timestamps.toDotNetJson(now);
+        var res = Timestamps.parseDotNetJson(jsonTimestamp);
         asr.equal(now, res);
     }
 
 
     test("DateTimes.MS_PER_DAY", function MS_PER_DAYTest() {
-        asr.equal(DateTimes.MS_PER_DAY, 24 * 60 * 60 * 1000);
+        asr.equal(DateConstants.MS_PER_DAY, 24 * 60 * 60 * 1000);
     });
 
 
@@ -43,28 +45,28 @@ suite("DateTimes", function DateTimesTest() {
 
     test("Dates.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
-        var res = DateTimes.Dates.getDayMinutes(date);
+        var res = Dates.getDayMinutes(date);
         asr.equal(134, res);
     });
 
 
     test("Dates.toDisplayDate", function toDisplayDateTest() {
-        var res1 = DateTimes.Dates.toDisplayDate(new Date(2000, 0, 27, 2, 14));
+        var res1 = Dates.toDisplayDate(new Date(2000, 0, 27, 2, 14));
         asr.equal("01/27/2000", res1);
 
-        var res2 = DateTimes.Dates.toDisplayDate(new Date(2000, 0, 9, 2, 14), "-");
+        var res2 = Dates.toDisplayDate(new Date(2000, 0, 9, 2, 14), "-");
         asr.equal("01-09-2000", res2);
     });
 
 
     test("Dates.toDisplayDateTime", function toDisplayDateTimeTest() {
-        var res1 = DateTimes.Dates.toDisplayDateTime(new Date(2000, 0, 27, 0, 0), false);
+        var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 0, 0), false);
         asr.equal("01/27/2000", res1);
 
-        var res1 = DateTimes.Dates.toDisplayDateTime(new Date(2000, 0, 27, 2, 14), false);
+        var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 2, 14), false);
         asr.equal("01/27/2000 02:14 a.m.", res1);
 
-        var res1 = DateTimes.Dates.toDisplayDateTime(new Date(2000, 0, 27, 23, 8), false);
+        var res1 = Dates.toDisplayDateTime(new Date(2000, 0, 27, 23, 8), false);
         asr.equal("01/27/2000 11:08 p.m.", res1);
     });
 
@@ -72,27 +74,27 @@ suite("DateTimes", function DateTimesTest() {
     test("Dates.dayDiff", function dayDiffTest() {
         var oldDate = new Date(2000, 0, 27);
         var newDate = new Date(2000, 2, 1); // Jan 27 to Mar 1 = 34 days
-        var res1 = DateTimes.Dates.dayDiff(newDate, oldDate);
+        var res1 = Dates.dayDiff(newDate, oldDate);
         asr.equal(res1, 34);
     });
 
 
     test("Timestamp.currentTimezoneOffsetMillis", function currentTimezoneOffsetMillisTest() {
-        var res1 = DateTimes.Timestamp.currentTimezoneOffsetMillis;
+        var res1 = DateConstants.currentTimezoneOffsetMillis;
         asr.equal(res1, new Date().getTimezoneOffset() * 60 * 1000);
     });
 
 
     test("Timestamp.now", function nowTest() {
         var now = new Date().getTime();
-        var res1 = DateTimes.Timestamp.now();
+        var res1 = Timestamps.now();
         asr.equal(res1 >= now, true);
     });
 
 
     test("Timestamp.toDate", function toDateTest() {
-        var now = DateTimes.Timestamp.now();
-        var res1 = DateTimes.Timestamp.toDate(now);
+        var now = Timestamps.now();
+        var res1 = Timestamps.toDate(now);
         asr.equal(res1.getTime(), now);
     });
 
@@ -114,28 +116,28 @@ suite("DateTimes", function DateTimesTest() {
 
     test("Timestamp.getDayMinutes", function getDayMinutesTest() {
         var date = new Date(2000, 0, 27, 2, 14);
-        var res = DateTimes.Timestamp.getDayMinutes(date.getTime());
+        var res = Timestamps.getDayMinutes(date.getTime());
         asr.equal(134, res);
     });
 
 
     test("Timestamp.toDisplayDate", function toDisplayDateTest() {
-        var res1 = DateTimes.Timestamp.toDisplayDate(new Date(2000, 0, 27, 2, 14).getTime());
+        var res1 = Timestamps.toDisplayDate(new Date(2000, 0, 27, 2, 14).getTime());
         asr.equal("01/27/2000", res1);
 
-        var res2 = DateTimes.Timestamp.toDisplayDate(new Date(2000, 0, 9, 2, 14).getTime(), "-");
+        var res2 = Timestamps.toDisplayDate(new Date(2000, 0, 9, 2, 14).getTime(), "-");
         asr.equal("01-09-2000", res2);
     });
 
 
     test("Timestamp.toDisplayDateTime", function toDisplayDateTimeTest() {
-        var res1 = DateTimes.Timestamp.toDisplayDateTime(new Date(2000, 0, 27, 0, 0).getTime(), false);
+        var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 0, 0).getTime(), false);
         asr.equal("01/27/2000", res1);
 
-        var res1 = DateTimes.Timestamp.toDisplayDateTime(new Date(2000, 0, 27, 2, 14).getTime(), false);
+        var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 2, 14).getTime(), false);
         asr.equal("01/27/2000 02:14 a.m.", res1);
 
-        var res1 = DateTimes.Timestamp.toDisplayDateTime(new Date(2000, 0, 27, 23, 8).getTime(), false);
+        var res1 = Timestamps.toDisplayDateTime(new Date(2000, 0, 27, 23, 8).getTime(), false);
         asr.equal("01/27/2000 11:08 p.m.", res1);
     });